started working on the horizontal slider
There's still a lot of work to be done in the horizontal animation,
but it kind of works now. I have to fix what happens when it reaches
the end. For now the base animation and the horizontal animation are
in different branches because it was easier to work on them
separately.